Extract HTTP mock step and infrastructure in own module.
This not really a common feature, so let's keep it on the side.
Plus I would like to migrate the server away from akka-http to something minimal in order to stop the propagation the ActorSystem around the code base.
