+++ cornichon-experimental/src/main/scala/com/github/agourlay/cornichon/sbtinterface/CornichonRunner.scala
@@ -8,7 +8,7 @@ import sbt.testing._
 import scala.concurrent.Await
 import scala.concurrent.duration.Duration
 
-class CornichonRunner(val args: Array[String], val remoteArgs: Array[String], testClassLoader: ClassLoader) extends Runner {
+class CornichonRunner(val args: Array[String], val remoteArgs: Array[String]) extends Runner {
 
   private val gotTasks = new AtomicBoolean(false)
