+++ cornichon-http-mock/src/main/scala/com/github/agourlay/cornichon/http/server/MockServerResultsHolder.scala
